+++ NEWS.md
@@ -1,5 +1,6 @@
 # frictionless (development version)
 
+* `read_resource()` now returns error if both `path` and `data` are provided (#143).
 * [checklist](https://github.com/inbo/checklist) tooling was removed, in favour of `CITATION.cff` for citation and Zenodo deposit.
 * `write_package()` no longer writes to "." by default, since this is not allowed by CRAN policies. The user needs to explicitly define a directory (#205).
 * `write_package()` encodes NULL and NA values as NULL (#203).